Added user roles to users. Content managers can now register accounts and create providers.

+++ FSE_Subscription_App/FSE_Subscription_App/Controllers/ContentController.cs
@@ -10,13 +10,13 @@ using System.IO;
 
 namespace FSE_Subscription_App.Controllers
 {
+	[Authorize]
     public class ContentController : Controller
     {
         private AppDbContext db = new AppDbContext();
 
         //
         // GET: /Content/
-
         public ActionResult Index()
         {
             var content = db.Content.Include(c => c.Provider);
@@ -25,7 +25,6 @@ namespace FSE_Subscription_App.Controllers
 
         //
         // GET: /Content/Details/5
-
         public ActionResult Details(int id = 0)
         {
             Content content = db.Content.Find(id);
@@ -35,7 +34,7 @@ namespace FSE_Subscription_App.Controllers
             }
             return View(content);
         }
-
+		
 		public FilePathResult ViewContent(int id = 0)
 		{
 			Content content = db.Content.Find(id);
@@ -48,7 +47,7 @@ namespace FSE_Subscription_App.Controllers
 
         //
         // GET: /Content/Create
-
+		[Authorize(Roles = "ContentManager")]
         public ActionResult Create()
         {
             ViewBag.ProviderID = new SelectList(db.Providers, "ID", "CompanyName");
@@ -60,6 +59,7 @@ namespace FSE_Subscription_App.Controllers
 
 		[HttpPost]
 		[ValidateAntiForgeryToken]
+		[Authorize(Roles = "ContentManager")]
 		public ActionResult Create(Content content, HttpPostedFileBase file)
 		{
 			if (ModelState.IsValid)
@@ -93,7 +93,7 @@ namespace FSE_Subscription_App.Controllers
 
         //
         // GET: /Content/Edit/5
-
+		[Authorize(Roles = "ContentManager")]
         public ActionResult Edit(int id = 0)
         {
             Content content = db.Content.Find(id);
@@ -110,6 +110,7 @@ namespace FSE_Subscription_App.Controllers
 
         [HttpPost]
         [ValidateAntiForgeryToken]
+		[Authorize(Roles = "ContentManager")]
         public ActionResult Edit(Content content)
         {
             if (ModelState.IsValid)
@@ -144,7 +145,8 @@ namespace FSE_Subscription_App.Controllers
         {
             Content content = db.Content.Find(id);
 			string filePath = content.ServerPath;
-			System.IO.File.Delete(filePath);
+			if(System.IO.File.Exists(filePath))
+				System.IO.File.Delete(filePath);
             db.Content.Remove(content);
             db.SaveChanges();
             return RedirectToAction("Index");
